Feature ETP-3799: Address Copilot review: deduplicate Create Contact button
- Extract createBtn local var in SearchInput to replace 3 identical
  inline button JSX blocks (DUPLICATED_BLOCK finding)
- Extract useCreateContactModal hook from sales/purchase order windows
  to eliminate the duplicate bpApiBaseUrl + createContactCtxValue
  useMemos (DUPLICATED_BLOCK finding)
